Add resources to exec
