+++ src/controller_examples/rabbitmq_controller/exec/mod.rs
@@ -1,4 +1,5 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 pub mod reconciler;
+pub mod resource;
 pub mod types;
