+++ src/controller_examples/rabbitmq_controller/spec/resource/mod.rs
@@ -1,0 +1,2 @@
+// Copyright 2022 VMware, Inc.
+// SPDX-License-Identifier: MIT
